Refactor key management and auto-id utils
- Moved environment loading and key registration functions to a new `utils` module.
- Updated all relevant files to import these utilities from the new module.
- Added `subjectCommonName` parameter to `registerIssuerAutoId` and `registerLeafAutoId` functions for better flexibility.
- Modified `revokeCertificate` and `deactivateAutoId` methods in `Registry` to accept `filePath` for signing.
- Cleaned up redundant code and improved code readability.
