+++ examples/node/src/auto-id/revoke-self.ts
@@ -2,44 +2,10 @@
  * Revoke self/issuer certificate
  */
 
-import { CertificateManager, Registry, generateEd25519KeyPair, saveKey } from '@autonomys/auto-id'
+import { Registry } from '@autonomys/auto-id'
 import { Keyring } from '@polkadot/api'
 import { cryptoWaitReady } from '@polkadot/util-crypto'
-import { config } from 'dotenv'
-
-function loadEnv(): { RPC_URL: string; KEYPAIR_URI: string } {
-  const myEnv = config()
-  if (myEnv.error) {
-    throw new Error('Failed to load the .env file.')
-  }
-
-  const RPC_URL = process.env.RPC_URL
-  if (!RPC_URL) {
-    throw new Error('Please set your rpc url in a .env file')
-  }
-
-  const KEYPAIR_URI = process.env.KEYPAIR_URI
-  if (!KEYPAIR_URI) {
-    throw new Error('Please set your keypair uri in a .env file')
-  }
-
-  return { RPC_URL, KEYPAIR_URI }
-}
-
-async function registerAutoId(registry: Registry, filePath: string): Promise<string> {
-  const issuerKeys = await generateEd25519KeyPair() // Ed25519
-  saveKey(issuerKeys[0], filePath)
-
-  const selfIssuedCm = new CertificateManager(null, issuerKeys[0], issuerKeys[1])
-  const selfIssuedCert = await selfIssuedCm.selfIssueCertificate('test200')
-  const registerIssuer = await registry.registerAutoId(selfIssuedCert)
-  CertificateManager.prettyPrintCertificate(selfIssuedCert)
-  const issuerAutoIdIdentifier = registerIssuer.identifier!
-  console.log(
-    `===\nRegistered auto id from issuer cert: ${CertificateManager.getCertificateAutoId(selfIssuedCert)} with identifier: ${issuerAutoIdIdentifier} in block #${registerIssuer.receipt?.blockNumber?.toString()}`,
-  )
-  return issuerAutoIdIdentifier
-}
+import { loadEnv, registerIssuerAutoId } from './utils'
 
 async function main() {
   await cryptoWaitReady()
@@ -55,10 +21,15 @@ async function main() {
 
   /* Register Auto ID */
   const filePath = './res/private.issuer.pem'
-  const issuerAutoIdIdentifier = await registerAutoId(registry, filePath)
+  const issuerSubjectCommonName = 'test200'
+  const [issuerAutoIdIdentifier, _issuerCm] = await registerIssuerAutoId(
+    registry,
+    filePath,
+    issuerSubjectCommonName,
+  )
 
   /* Revoke self Certificate */
-  const revoked = await registry.revokeCertificate(issuerAutoIdIdentifier)
+  const revoked = await registry.revokeCertificate(issuerAutoIdIdentifier, filePath)
   if (revoked) {
     console.log(
       `Revoked registered issuer certificate with identifier: ${issuerAutoIdIdentifier} in block #${revoked.blockNumber?.toString()}`,
